Design: Community UI 커뮤니티 상세 페이지 UI 수정 - CircularAvatar 추가 - 이름, 날짜 배치 변경 - textSize, padding, margin 조정 - Tag 추가 댓글 수정 및 삭제 버튼 imageButton -> menu option으로 변경 권한 있는 사용자만 수정, 삭제 가능 findDifference 함수 Utility로 이동(재사용 위해)
