+++ app/src/main/res/menu/comments_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/option_edit"
+        android:title="댓글 수정하기" />
+    <item
+        android:id="@+id/option_delete"
+        android:title="댓글 삭제하기" />
+</menu>
